fixed bug that prevented warcs bigger than 2 GiB to be processed due to max integer size

+++ src/main/java/pt/arquivo/imagesearch/indexing/ImageIndexerWithDupsJob.java
@@ -195,7 +195,7 @@ public class ImageIndexerWithDupsJob extends Configured implements Tool {
                 String filename = warcFileTempBaseDir + "/" + System.currentTimeMillis() + "_" + arcName;
 
                 try {
-                    int fileSize = ImageSearchIndexingUtil.getFileSize(url);
+                    long fileSize = ImageSearchIndexingUtil.getFileSize(url);
                     //ImageSearchIndexingUtil.saveFile(url, filename);
                     File dest = new File(filename);
